Test www.mortgagecalculator.org and www.flickr.com
Part 1 utilizes Selenium Web Driver to navigate
www.mortgagecalculator.org, identify and act on Web Elements, such as
checked boxes, buttons, text input boxes, and etc. At the end, it
identify certain texts on the web page, store them in the program, and
print them to the console. The driver also takes screenshots of before
and after calculation.

Part 2: The program navigate to the Explore page of www.flickr.com.
Download all the photos and store them to a designated folder.